Revert "Merge pull request #4691 from davecheney/resource-api-move-apiserver-common-restoreerror"
This reverts commit cc383e1beb1d8335a52aa207db6b20802a9bc053, reversing
changes made to ae58617b782f7acad3b3950a12b01c2816cec8c7.
