+++ payload/api/private/helpers.go
@@ -14,7 +14,6 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/payload"
 	"github.com/juju/juju/payload/api"
-	resourcesapi "github.com/juju/juju/resource/api"
 )
 
 // NewPayloadResult builds a new PayloadResult from the provided tag
@@ -50,7 +49,7 @@ func API2Result(r PayloadResult) (payload.Result, error) {
 	}
 
 	if r.Error != nil {
-		result.Error = resourcesapi.RestoreError(r.Error)
+		result.Error = common.RestoreError(r.Error)
 	}
 
 	return result, nil
