+++ resource/api/private/client/client.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/juju/errors"
 
+	"github.com/juju/juju/apiserver/common"
 	"github.com/juju/juju/resource"
 	"github.com/juju/juju/resource/api"
 	"github.com/juju/juju/resource/api/private"
@@ -87,7 +88,7 @@ func (c *UnitFacadeClient) getResourceInfo(resourceName string) (resource.Resour
 		return resource.Resource{}, errors.Annotate(err, "could not get resource info")
 	}
 	if response.Error != nil {
-		err := api.RestoreError(response.Error)
+		err := common.RestoreError(response.Error)
 		return resource.Resource{}, errors.Annotate(err, "request failed on server")
 	}
 
@@ -95,7 +96,7 @@ func (c *UnitFacadeClient) getResourceInfo(resourceName string) (resource.Resour
 		return resource.Resource{}, errors.New("got bad response from API server")
 	}
 	if response.Resources[0].Error != nil {
-		err := api.RestoreError(response.Error)
+		err := common.RestoreError(response.Error)
 		return resource.Resource{}, errors.Annotate(err, "request failed for resource")
 	}
 	res, err := api.API2Resource(response.Resources[0].Resource)
